added firebase to project
