--- app/src/main/java/com/mobolajia/notesapp/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.mobolajia.notesapp
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
