+++ app/src/main/java/com/mobolajia/notesapp/ui/LoginActivity.kt
@@ -1,0 +1,29 @@
+package com.mobolajia.notesapp.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import com.mobolajia.notesapp.R
+
+class LoginActivity : AppCompatActivity() {
+
+    private lateinit var auth : FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+
+        auth = Firebase.auth
+    }
+
+    override fun onStart() {
+        super.onStart()
+
+        val curUser = auth.currentUser
+        if (curUser != null) {
+            //do logout
+        }
+    }
+}
