+++ app/src/main/java/com/mobolajia/notesapp/ui/MainActivity.kt
@@ -1,11 +1,12 @@
-package com.mobolajia.notesapp
+package com.mobolajia.notesapp.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.mobolajia.notesapp.R
 
-class RegisterActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register)
+        setContentView(R.layout.activity_main)
     }
 }
