+++ app/src/main/java/com/mobolajia/notesapp/ui/RegisterActivity.kt
@@ -1,0 +1,37 @@
+package com.mobolajia.notesapp.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import com.mobolajia.notesapp.R
+import com.mobolajia.notesapp.databinding.ActivityRegisterBinding
+
+class RegisterActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityRegisterBinding
+    private lateinit var auth : FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        auth = Firebase.auth
+    }
+
+    private fun registerNewUser(email : String, password : String) {
+        if (email.isNotEmpty()) {
+            if (password.isNotEmpty()) {
+                auth.createUserWithEmailAndPassword(email, password)
+                    .addOnCompleteListener {
+                        if (it.isSuccessful) {
+
+                        } else if (it.isCanceled) {
+
+                        }
+                    }
+            }
+        }
+    }
+}
