+++ app/src/main/java/com/mobolajia/notesapp/viewmodel/LoginViewModel.kt
@@ -1,0 +1,6 @@
+package com.mobolajia.notesapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class LoginViewModel : ViewModel() {
+}
