+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="24dp"
-    tools:context=".ForgotPasswordActivity">
+    tools:context=".ui.ForgotPasswordActivity">
 
     <TextView
         android:id="@+id/forgot_password_txt"
