+++ app/src/main/res/layout/activity_register.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:padding="24dp"
-        tools:context=".RegisterActivity">
+        tools:context=".ui.RegisterActivity">
 
         <TextView
             android:id="@+id/register_txt"
